Restore the smokey_deploy job in AWS Staging and AWS Production
This job was removed in [1], but the smokey-loop serivce is still
running on the monitoring boxes in both environments.

While there may be some problems with smokey in AWS environments, it's
still useful to be able to run the deploy Jenkins job, so that at
least an up to date version is running.

1: ce9f0051dcfd8f43d6b7324cdd72f0b353ad7d7f