Reduce monitoring disk warning threshold to 12%
This was reduced to 15% in #8314, to handle the increased disk usage
due to the CDN logs.  However, we're still frequently going past that
threshold towards the end of each log rotation period (hourly).  I
haven't seen it drop below 13% free, so set 12% as the threshold.

This does mean that there's only 2 percentage points between the
warning and the critical thresholds, which seems not great though.