Trello-1728: Add mongo-api cluster to aws staging
We need to create a mongo cluster (mongo_api) that will be used for testing addition and removal on mongo cluster nodes which will be carried out in card #1725

The mongo cluster will be built using the same module as existing mongo
module: mongodb::server

If our tests are successful in #1725 we will be able to have a no-downtime migration of the mongo databases.
It is also envisaged that the mongo cluster will be split up, in other
words the content-store will use mongo-api cluster (that we are
currently building here) and immenence will continue to use the existing
mongo cluster.

There is hieradata being added to the carrenza side as part of this
commit, that is to satisfy the spec tests, once migrations are fully
completed this hieradata will be collapsed into one, so it is
temporary.

Pair: @th31nitiate @ronocg