Fix crawlers to crawl their own environment
Previously the crawlers for each environment were all crawling
www.gov.uk, which has recently (for some reason) generated rate limit
alerts, as the limit bypass tokens aren't shared between environments.

This changes the crawler config for staging (and similarly for AWS)
to crawl origin for each environment. We've chosen to crawl the origin
instead of the CDN in these environments because:

   * The Carenza machines map www -> www-origin anyway
   * The AWS mirror machines can't access the CDN
   * The load impact of the crawler isn't such an issue

Ideally we should crawl the CDN on all environments, using a common URL
structure (compare www.gov.uk vs www.staging.publishing...), but this
isn't essential, although we should still aim to do this for AWS.
Note that the crawler is disabled for integration (for some reason).