Stop crawling www-origin on staging (bad sitemap)
Previously we fixed the GOV.UK crawler in staging to crawl the staging
environment and not production. The intention was to use www-origin to
avoid confusion around the www -> www-origin mapping in /etc/hosts and
to better support the migration of this service to AWS.

This removes -origin from the config, as these are not compatible
with the sitemap.xml entries the crawler uses to do its job. This means
we will have to rely on the www -> www-origin mapping on Carrenza for
the time being (staging machines don't have access to the real www CDN).