Remove Whitehall Import job for Content Publisher
Previously we configured a Jenkins job on AWS Integration to
periodically import data from Whitehall into Content Publisher.

As we start to give integration access to our private beta partners, we
want to temporarily stop this import to avoid confusion.

Removing the Puppet configuration here will allow us to manually
disable/enable the job while we are in this transitional period.