Add alert app domain management subdomain to /etc/hosts
Trying to access Icinga alerts from Jenkins leads to
connection errors because there are no approriate hosts
defined which connect to the right server. Using the external
domain right now doesn't map to an internal IP address.

We can't use alert.cluster here because the HTTPS server is
checking that the domain name matches the server name in the
certificate.