Trello-1731-1774 - push/pull for api-mongo migr
This gets the push / pull mongo sync scripts and cron jobs in place for
the draft-content-store and the content-store miration to AWS.

The draft-content-store and the content-store both reside on a mongo
cluster called api-mongo in carrenza. As part of the migration both
these databases will move to an existing cluster in AWS and sit
alongside other mongo databases, that clustser's nodes are named mongo.

- the existing push cron jobs are being removed from api-mongo-4 in
  carrenza but the scripts and configfile for the syncs will now reside
  on all api-mongo nodes in carrenza.
- the pull scripts and config files will be laid down on the AWS side
  but will not have cronjobs associated with them
- the push scripts will be scheduled via cron on the AWS side

staging and production environments will have their push and pull
scripts setup in a similar manner.

Solo: @ronocg <conor.glynn@digital.cabinet-office.gov.uk>