Migr-plan: etc hosts for content-stode migration
Here we are removing the /etc/hosts entries for the content-store and
teh draft-content-store in Carrenza Staging. This is part of the
migration plan, so the publishing-api and the frontends in carrenza
staging will begin communicating with the content-store and
draft-content-store in aws.

also removing a swp file that was committed at some stage.

Pair: @th31nitiate <david.jeche@digital.cabinet-office.gov.uk>
& @ronocg <conor.glynn@digital.cabinet-office.gov.uk>