Trello-359: Staging push sync
There are 13 push data sync jobs in staging that are currently
Acknowledged as failing on icinga. These push jobs are not required in
staging any longer and can be disabled to clear the icinga alerts.
During migration of these databases to AWS Staging we will be
temporarily enabling the jobs again to do a push during the migration.

Solo: @ronocg <conor.glynn@digital.cabinet-office.gov.uk>