Add monitoring for the cache-clearing-service queue size
The cache clearing service reads from a RabbitMQ queue, which has
peaked at ~1,000,000 items recently. This commit adds some monitoring
via Icinga so that it's easier to see when the queue is "long".

The values are a complete guess, at the moment, it seems to be backing
up significantly, so hopefully this can be solved, and then the alert
values tuned to a more sensible range.