Indicate pending items in a terraform plan
# Context

The deploy terrafrom govuk aws project allows for plans and applies of
terraform but when running a plan there is no indicator that there are pending
changes. This can be misleading if scanning down a list of "green" jobs, the
assumption could be that everything is working and deployed.

# Decision

Use the Jenkins Text Finder plugin in a post-build action to identify pending
changes and mark the job as unstable (yellow).