Mock the operatingsystem fact in phantomjs spec tests
The phantomjs module tries to install a specific version of phantomjs. This
requires a package provider which is 'versionable', which is not the case when
running the tests on OS X. Pretend we're on Linux instead, which is enough to
trick puppet into picking a 'versionable' provider for package resources.