Increase the high CPU thresholds for Asset Manager
When large numbers of documents are uploaded to Asset Manager, it
has to run the virus scanner on each of those. This causes brief
spikes in CPU usage.

This has been the case since around September 2018, especially
since Whitehall was migrated to use Asset Manager for attachments.
I think we can probably increase the thresholds since it doesn't
appear to be causing any significant issues, and we have other
alerts that will tell us if Asset Manager is starting to serve 5xx
errors which could be a consequence of high CPU usage.