Indicate that planned items exist
# Context

When running terraform plan there is no indicator that items need to be
applied. This could lead to undesirable assumptions, namely that changes are
deployed when they aren't.

# Decision

Using Jenkins Text Finder it's possible to change the state of a job to
unstable based on a regex match. The regex in this PR will only appear if there
are planned items, if there is nothing to apply the regex will not match
anything.