Change the notification_period for the Publishing API healthcheck
This healthcheck was added following some incidents with delayed
publishing due to issues with the Publishing API downstream high
queue.

The healthcheck now includes the queue latency. The check was first
added using the inoffice notification period, for testing purposes,
but now that it hasn't alerted spuriously, it's being changed to 24x7,
as this is something that needs to work all day, every day.