Revert "disable smokey tests in staging and production"
I don't think this had the intended effect, as I believe Smokey tests
are still enabled in Staging and Production.

Furthermore, by re-enabling deployements, and improving
Smokey (e.g. [1]), it seems to be more reliable now.

Currently, the smokey-loop service in AWS Staging is configured to use
the production profile (the default), due to the value in the
hieradata being commented out. But the profile doesn't really have an
effect at the moment, so this change shouldn't either.

1: https://github.com/alphagov/smokey/pull/438

This reverts commit cec39f363fd8ff1dff0eb299ca8cc7a11e031c4d and
commit 9cfe01a975253ddd210aa306db21b7efd2ebeb0c.