Temporarily disable the CPM daily push
The configuration/sync script isn't currently able to handle the
Content Performance Manager database, so the push is failing. I'm not
sure how to fix it, so for now, just disable it.

There seems to be multiple issues, both around the hostname the data
is read from, and the PostgreSQL user used for the connection.