Revert "remove the smokey and deploy smokey jobs"
The Smokey deploy job has already been restored, and this commit
restores the Smokey job as well.

The Smokey Jenkins job is triggered when deploying apps, and as we are
now running apps in AWS based environments, it's important to be
running Smokey as part of the deployement process, to help identify
issues.

This was disabled as Smokey wasn't running reliably, but it now runs
reliably on the monitoring boxes, so I see no reason why we shouldn't
try it again on the deploy machines.

This reverts commit ce9f0051dcfd8f43d6b7324cdd72f0b353ad7d7f.