Improve the event-store configuration
Transform the enabled value in to the ensure value, and then pass that
value in to the different components, including the environment
variables.

Then, disable the app in both the AWS and non-AWS hieradata, removing
the redundant value from the development specific hieradata.

These changes should solve some govuk-puppet errors related to the
event-store.