Remove conditional around include loadbalancer::maintenance
The sole purpose of this configuration is to ensure the loadbalancer
has the page to return when in maintainance mode.

It's best to include it all the time, so all of the loadbalancers are
ready for maintainance mode, and so the relevant code is used and
tested, before you actually need it.