Check that AWS and Carrenza hieradata are in sync
We currently have to duplicate the AWS and Carrenza hieradata. This adds a test to check that they have the same keys defined. To account for legitimate differences, we also implement two whitelists.

We currently have a *lot* of differences, so I've whitelisted everything for now. This will at least stop us from drifting further apart. We'll need to manually whitelist & fix any existing discrepancies.