Remove nginx::restart
We already define a restart command on the 'nginx' service, so we can replace
this class with a notification against the service.

Use a resource collector in nginx::package, because that doesn't include the
service (so it may not be defined)

Sprinkle some code and test tidyups around whilst we're at it.