Add dummy search-api configuration to hierdata/common.yaml
All keys present in hieradata_aws/common.yaml need to be present in
hieradata/common.yaml (and vice versa).  So include some dummy values
here to make CI happy - their nonsense values are fine because
search-api isn't in the list of deployable apps for Carrenza.

In addition, the dev VM uses this configuration, so the configuration
has to be specified for that to build.  Getting this actually working
in the dev VM is a challenge for another day.