Ensure the workspace is clear before running the job
# Context
When running this job there is a java.nio.file.FileSystemException error when attempting to remove the git repo.

# Decision
Clear the workspace prior to running the job. This was though to have been done in a previous commit but the yaml there was wrong. Adding the wrapper is the correct way to achieve this.