Trello-1921: nginx loadbalancer config
As part of the migration of the frontends to aws, there needs to be a
way for the aws finder-frontend to access the search-api (search-1/2/3
VMs) in carrenza. While in carrenza this was accessed using an
internal-lb via a hosts entry but aws will dns resolve
search.publishing.service.gov.uk which will be cnamed into the
backend-lb.

Therefore the backend-lb will require the nginx configuration for search
which is being laid down by this PR.

ALso limiting the nginx for search to only allow from the 3 nat gateway
egress ips from the corresponding AWS environment.

@ronocg <conor.glynn@digital.cabinet-office.gov.uk>