Increase timeout for Content Data Admin
We are seeing timeout issues when the Content Performance Manager
is running slow queries. The timeout does not occur at the API end
and we are not able to reproduce this error on other environments.
This problem only occurs on production. For now we are upping the
timeout on the client end to address the timeouts, and allow
opportunities for performance optimisation later.