Increase cache-clearing-service RabbitMQ thresholds
Since we made the queue automatically expire messages after an
hour, there's less of a risk of this growth being a significant
problem.

I've looked back in the last week and the highest the queue ever got
to was 80k, so it seems fine to increase the thresholds to those
sorts of levels.