Create elasticsearch plugin .name files
The version of puppet-elasticsearch we're upgrading to in #8594
expects these files to exist, and uses them in its plugin version
detection.  If the plugin is installed but without the .name file (as
is currently the case), it will give an error asking you to manually
uninstall the plugins first.

This doesn't seem to be mentioned in the CHANGELOG with the other
breaking changes.