Increase total process limits
We're starting to see this flag up in Icinga because each
monitoring process is a single process and as we add more monitoring
this is likely to hit the limit.

Unfortunately this meaning increasing the global thresholds, but in
reality on most machines the process count isn't anywhere near the
levels right now and this alert is only relevant for the monitoring
machines.