Schedule tag_metadata Rummager rake task twice daily
There's a bug in how updates to the mainstream index write metadata
via the Indexer::MetadataTagger class in Rummager which surfaces
when Publisher content updates are made to content present in the
'Find EU Exit guidance for your business' finder.
Add a job to schedule running the retagging rake task which successfully
re-applies the relevant metadata, this is scheduled in line with 2ndline
timing for applying updates for convenience.