Trello-1963: Envars for the frontend
The PLEK envars for teh draft_frontend are laid down in
modules/govuk/manifests/node/s_draft_frontend.pp  but becuase that class
includes the s_frontend class there would be duplication of the envars
in puppet. Here we need to make the if conditional explicit to just
manage the vars if the node is a frontend noe in aws staging or
production.

@ronocg <conor.glynn@digital.cabinet-office.gov.uk>