Staging-Migrations-Fix: Plek Envvar required
smartanswers frontend needs to access whitehall-admin (which is staying
in carraenza for the moment).
Currently smartanswers in the AWS environments attempts to access
whitehall-admin using the govuk.digital name. So in order to override
that we are laying down a PLEK envvar on the calculators-frontend
machines for whitehall-admin so traffic will be directed to carrenza
whitehall-admin..

@ronocg <conor.glynn@digital.cabinet-office.gov.uk>