Trello-2020: Plek envvars for whitehall on content-store and draft-content-store
When router-api and its database migrate to AWS, whitehall is staying in
carrenza, therefore the whitehall backends in the router-api DB will need
to have the associated publishing.service.gov.uk domain.
This has been completed already by manually setting the whitehall backend
in the router api database with the correct domain.
BUT these could possibly be updated at any time from content-store, so
content-store and draft-content-store need the PLEK envvar in place for
the whitehall-frontend URI.

The end state for the variable should be;

staging:
https://whitehall-frontend.staging.publishing.service.gov.uk

production:
https://whitehall-frontend.publishing.service.gov.uk

Also because this parameter will be undef for other environments other than
aws staging and aws production they will not be laid down for those
other envs.

@ronocg <conor.glynn@digital.cabinet-office.gov.uk>
@jstandring-gds <julian.standring@digital.cabinet-office.gov.uk>