Remove app_domain hieradata for content store forcing default
# Context

As part of the migration we need to update the app domain for the content store
so that the content store is aware of the new app (frontend) locations. This
was previously done by changing the value for
govuk::apps::content_store::app_domain, however if this data is absent then the
app_domain for content store is set to the appropriate default value.

# Decision

Remove the hiera data so that we fall back to the default defined in the
govuk::deploy::config class.