Change the contact_groups for some AWS machines
Specifically the content-store and backend machines. It's unnecessary
to have this severity of alerting for these machines, as the recovery
process is entirely automated.

For the backend, content-store and draft-content-store machines,
requests are load balanced, and when a machine goes away, requests
will eventually stop being sent to it, and a new machine will be
brought up to replace it.

If anything goes wrong with these machines and it is necessary for
someone to intervene out of hours, that should be detected by alerting
on the important thing that's going wrong, e.g. too many requests with
a 5xx status reaching users.