Temporarily disable the staging data sync for staging
We need to run a full reindex of Solr at least once to be able to
run the data sync process which only indexes changes documents.

Unfortunately this full reindex takes more than 3 days which means
when the data sync runs the process stops because the database is
temporarily unavailable. So we should disable the data sync until
we can run the full reindex first.