Add Puppet configuration for the Content Data API
The Content Performance Manager application is being renamed to the
Content Data API to better reflect it's current function.

The Puppet configuration added in this commit relates to the new
configuration for the Content Data API. It's basically a copy of the
Content Performance Manager configuration.

Some things are staying the same for now, like the repository and
database. I think it'll be smoother to change these things once the
app is up and running under the Content Data API name.

Once the migration from the Content Performance Manager to the Content
Data API has been completed, the Content Performance Manager
configuration can be removed.