Raise memory thresholds for static
Static seems to be settling at a level just below the warning line.  Last night
this resulted in a restart cycle (due to exceeding the critical level) that took
a while to settle down.

We're generally more relaxed about having strict thresholds on memory thresholds
these days as we have alerts on whether apps are actually working.

There's not a massive amount of free memory on these machines, but we'll be moving
to AWS very soon.