Carrenza Icinga reboot alert shouldn't retrieve puppet class
In Carrenza, the govuk_node_list does not have the --with-puppet-class
option which was only added to AWS govuk_node_list because AWS
hostnames do not contain the puppet class and therefore, this new
option was added in AWS.

Hence, the Icinga reboot alert should not use the --with-puppet-class
in Carrenza. In any case, the hostname in Carrenza already contains
the puppet class by default.