Update Production app domain for migration
# Context

As part of the migration we need to update the app domain for the content store
so that the content store is aware of the new app (frontend) locations.

# Decision
Remove the hiera data so that we fall back to the default defined in the
govuk::deploy::config class.

This is a replay PR https://github.com/alphagov/govuk-puppet/pull/8679