Make Upstart more lenient to harvesting restarts
We often see the harvesting process crash due to a bug in the code.
After restarting, it will proceed with the next item in the
harvesting queue and continue normally. However, sometimes we see
that there are lots of harvesting jobs which cause the process to
crash grouped together in the queue. This causes Upstart to think
the harvesting process isn't working properly and stops restarting
it.

Until we have sufficient alerting about this process we can tell
Upstart to be more lenient when decided whether the process has
died or not and make it more likely to be restarted.