Remove package_ensure => absent from s_content_data_api_db_admin
Abusing the Puppet PostgreSQL module for RDS configuration requires
letting Puppet setup the DB Admin machine like it's the thing it's
configuring, but tricking it to connecting to the RDS instance for
some configuration.

I'm trying to avoid as much of this redundant and confusing setup on
the DB Admin machine as possible, but it's not going well. Puppet
seems to require the package being installed, as it's still trying to
ensure directories (like /var/lib/postgresql/9.3) exist, and trying
use the postgres user and group, and trying to change the
configuration.

This commit just changes the PostgreSQL server package to be
installed, which performs some of the setup the PostgreSQL Puppet
module implicitly depends on.