Add permanent redirect for crown logo to asset-manager
1stline use /static/gov.uk_logotype_crown.png in their zendesk
templates, but that URL only works because we keep all deployments of
assets from 'static' in Carrenza, even ancient things like these
unfingerprinted files.

To give the file a fixed location, it now lives in the asset-manager.
To keep historical emails working (and anything else which uses this
URL), we redirect the old path to the new one.