Adjust the Content Data API database configuration
Temporarily make the database name and user match up with the Content
Performance Manager. I'm hoping this will make it easier to sync the
Content Performance Manager database from where it currently resides,
to the new infrastructure for the Content Data API.

Once the new service is up and running, just prior to switching over
to it, we can rename the database, change the ownership, and remove
this configuration.