Move search spelling/benchmark jobs to AWS
Since we moved the cache machines (www-origin) to AWS, we've been
getting SSL errors when we try to query it from the Carrenza monitoring
machine, which is likely due to the /etc/hosts entries that are still
present on these machines. As these jobs are API-based, we should be
able to run them from the AWS production deploy machine instead.