Use origin for search benchmark jobs
It's no longer possible to use origin on Carrenza production, but nor is
it possible to use the internal domain on AWS production, as the Search
API is not fully deployed yet. This switches the domain for AWS to be
origin in all environments, at least until the internal domain works.