Update lots of gem versions in the Gemfile.lock
This was done with Bundler 1.17.3, but I've removed the BUNDLED WITH
section from the Gemfile.lock to avoid older versions of Bundler
finding out and complaining.

I also added `ruby '1.9.3'` to the Gemfile, but I haven't committed
that change, as it seems to break Puppet deploys in some
environments (at least Carrenza Staging [1]).

1:
/usr/share/puppet/production/releases/20190401145703/Gemfile:3:in `evaluate'
undefined method `ruby' for #<Bundler::Dsl:0x000000016e0310> (NoMethodError)