Add ensure as a parameter to the email_alert_service::rabbitmq class
Then pass the value down.

Currently, this class is active everywhere it's present, but as we're
now running govuk-puppet in Carrenza and AWS, with different sets of
services. This configuration doesn't work well in AWS, and leads to
spurious alerts. Therefore, add a parameter so that this configuration
can be disabled.