Switch the Content Data Admin to use the Content Data API
Rather than the Content Performance Manager. This is the new name for
the application which is being deployed to the AWS environments.

Configure it to use the Content Data API in AWS Production and AWS
Staging, as it's not currently being used in these two environments,
and this allows testing of the Content Data API.