Make the datascrubber absent in AWS Production and AWS Staging
It doesn't make sense to me to scrub data in either of these
environments, removing sensitive data has only ever been done for
Integration as far as I'm aware.

Also, I'm not aware that any of the relevant databases that could be
effected currently exist in either of these environments, and the
acknowledged alerts in Icinga seem to suggest this is the case as
well.