Set the memory thresholds for CKAN
Now that the collectd_process_regex is set to a working value, we're
getting a real value for the memory used by CKAN. Turns out this is
around 2GB at the moment. I don't know if this is usual, but it seems
OK, so set the thresholds for alerting just higher than the current
value.