Fix Icinga Graphite URLs in AWS
By using app_domain_internal we can't access it externally, but by using
app_domain we end up with the Carrenza version of Graphite. While we
have two versions of Graphite and Icinga, we need a custom configuration
for AWS. We can't easily use Hierdata here because this is a definition
and not a class meaning that we would need an entry each time it's used,
so I deciced instead to hard code the app domain since it's only
temporary.