Remove -1 from psql import to get import working
Postgres sync from production ckan postgres is failing which means that ckan staging is down slowing down deployments.

To get it working we removed -1 from the psql args which means that rather than a single transaction the import will be done over multiple transactions.

The impact of this is that it takes a little bit longer for the import and there is a potential for corrupt data if another process is running against the staging postgres.

It's probably ok as the sync happens overnight so there should be no processes running and its only staging.