Don't delay compression of Nginx logs
Our logs for Nginx tend to be fairly large, and I don't see a reason to
delay the compression of the logs (which means that the first rotation
of a log doesn't get compressed). I looked back in the history of
logrotate configuration and it looks like this was kept around because
it's the default configuration for Ubuntu.

Here is a current example of the sizes of our logs:

438M	lb-access.log
1.3G	lb-access.log.1