Set the maxsize for more logs
I've gone through each logrotate configuration and tried to think of a
good maxsize value to use. I've set the limits fairly strictly because
we're starting to see more and more issues where the log files are too
big and cause the apps to break.

Plus, our logs are exported to Logit which means they are retained for
much longer than on the box.