Bump search-api unicorn workers to 9 per node
We can currently handle 2 nodes * 6 workers simultaneous requests to
search-api (this is the same as rummager).  The AWS search machines
are currently using about 30% of each core and a little under half the
RAM, so that number can probably go up.

To avoid the risk of RAM exhaustion, we don't necessarily want to
double this to 12 without checking the effect of a smaller increase
first.