Remove empty .1 log files before rotating
There seems to be a problem with logrotate whereby it will refuse
to rotate the log files if an empty .1 log file exists. This leads to
increase disk space being used by log files and can cause the disk to
fill up.

This fix should mean that all empty .1 log files are deleted before
logrotate runs which should allow it to finish successfully.