Fix process count / error rate graphs
These graphs are not consistently shown at the moment because the panels
are too small for the data in the time span.

By combining the graphs, we make them visible again, but also get the benefit
of aligning them with the other full width graphs on the dashboard, making
it clearer when events correlate.

The error rate is set against the right axis because the scale of this
is different from the process count which tends to be between 0 and 10 (or
thereabouts) depending on the application.