Bump finder-frontend unicorn workers to 6
We're seeing finder-frontend timeouts, even of the healthcheck which is just a static response.  This suggests that finder-frontend can't respond in a timely manner to the volume of requests it's receiving.

Increasing the number of workers will increase the number of requests to search-api, which is a concern, but search-api is using more workers than rummager was, so it can cope with a bit more node.