Remove unused fields_prefix
This was used when we had some logstasher log fields prefixed with `@fields` (so
`@fields.duration`, `@fields.action`, @fields.controller` etc.), and some which
has no prefix.

We started fixing this and trying to align on having no prefix back in Jan 2018
(https://github.com/alphagov/govuk-puppet/pull/6658). This should now be
complete as of https://github.com/alphagov/search-api/pull/42, so we can
remove references to the prefix that allows us to manage the convergence.