Delete the tempdir on exit to avoid filling up space
Often we see a situation where the sync fails to run but leaves around
all the temporary files. This means that the second attempt of running
the sync also fails because there is no more space available for the
temp files. Instead, we can always delete the temp directory when the
sync script exits. These files aren't often used for debugging so it
shouldn't cause to many problems.