Add Icinga monitoring for established TCP connections to apps
This should flag when requests are potentially being queued for
processing at the point of reaching Unicorn (or other similar things).

We've had issues with this for Finder Frontend recently, and this
seems to be the way to monitor if there's insufficient workers to
process all the requests. I'm not exactly sure why it's the
ESTABLISHED metric, but this correlates with timeout issues.