fix puppet cronjob to avoid conflict with puppetserver restart
Puppet server restarts every 30 mins and if it is not fully functional before other nodes request their puppet manifest, the other nodes will show a puppet error in icinga. In order to solve this issue, we want to have a minimum delay of 5 minutes between when the puppetserver restarts and the first node requests its manifect from the puppet server.