Change to use the content_data_api_bearer_token
For the Content Data Admin. This follows on from the work to rename
the Content Performance Manager to the Content Data API.

For compatability, both environment variables are still set. The
CONTENT_PERFORMANCE_MANAGER_BEARER_TOKEN one can be removed once the
app is deployed using the new environment variable.